requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
pillow
dj-config-url
django-filter
drf-yasg
